[wip] Interface with SANA-FE message traces
Add a custom traffic manager to simulate the packets generated
by spiking architectures. Model the neuron processing pipeline
and use this to inject packets at the right times.

Track when cores are busy so that we correctly model the backpressure
The capability is supported by booksim2, but add tracking of receiving
core pipelines and prevent messages from being received if the core
is still processing a message

TODO: support packet queues for sending / receiving messages within
each core. Maybe this can just be supported by having output queues
on every link.

TODO: clean up the mess that is my implementation. Get rid of the
message received handling inside the traffic manager (this has to
be added to the queue implementation)