Fix bug with rx node backpressure and subnets
Change how I track the rx node buffers, and whether they are
available to receive flits. Although the previous implementation
was cleaner, it didn't work with multiple subnetworks (because
the receiving node was duplicated for each subnet, whereas it
should be shared).

Move to a hackier global model, which globally tracks each
node and uses the router id to index into the correct buffer.