Fix message parsing bug and credit handling bug
Message parsing was assuming the neuron group is numbered, whereas
it should be named (I used group numbers for v1, now its group names).

Fix bug with credit not being processed, exposed by a toy SNN with a
single spike (taking a few cycles to process). Make sure we wait
for credit to be processed in addition to messages being received before
skipping timesteps. This bug caused credit to never be processed and so
we got stuck in an infinite loop for some tiny SNN cornercases.