More hacks to finalize support for multiprocessing
Code now supports multiple concurrent Booksim runs from the same
SANA-FE process. This was done through a mix of hacks - a mutex
for the parser which is only used at initialization, and use of
thread_local for static variables and a global config state object
(replacing global variables). Not pretty, but functional for now.